scripts/requirements.txt
plotly==5.13.1
kaleido==0.2.1
python-dateutil==2.8.2